refactor(vm): separate stack and call mechanics
