+++ internal/vm/architecture_test.go
@@ -45,3 +45,13 @@ func TestBuiltinSourceLayout(t *testing.T) {
 		requireSourceFunctions(t, filename, names...)
 	}
 }
+
+func TestStackAndCallSourceLayout(t *testing.T) {
+	expected := map[string][]string{
+		"stack.go": {"readShort", "isFalsey", "valuesEqual", "readConstant", "push", "pop", "peek", "captureUpvalue", "closeUpvalue"},
+		"calls.go": {"callValue", "call", "copyValue"},
+	}
+	for filename, names := range expected {
+		requireSourceFunctions(t, filename, names...)
+	}
+}
