+++ internal/vm/vm.go
@@ -1431,223 +1431,6 @@ func (vm *VM) run(minFrameCount int) error {
 	}
 }
 
-func (vm *VM) callValue(callee value.Value, argCount int, c *chunk.Chunk, ip int) (bool, error) {
-	if callee.Type == value.VAL_OBJ {
-		if structDef, ok := callee.Obj.(*value.ObjStruct); ok && structDef != nil {
-			// Instantiate
-			if argCount != len(structDef.Fields) {
-				return false, vm.runtimeError(c, ip, "expected %d arguments for struct %s but got %d", len(structDef.Fields), structDef.Name, argCount)
-			}
-			args := vm.stack[vm.stackTop-argCount : vm.stackTop]
-			if err := vm.validateStructConstructorArguments(structDef, args); err != nil {
-				return false, vm.runtimeError(c, ip, "%s", err)
-			}
-
-			instance := value.NewInstance(structDef)
-			instObj := instance.Obj.(*value.ObjInstance)
-
-			// Args are on stack.
-			for i := 0; i < argCount; i++ {
-				arg := vm.peek(argCount - 1 - i)
-				fieldName := structDef.Fields[i]
-				instObj.Fields[fieldName] = arg
-			}
-
-			// Pop args AND callee (struct def)
-			vm.stackTop -= argCount + 1
-			// Push instance
-			vm.push(instance)
-			return true, nil
-		}
-	}
-	if callee.Type == value.VAL_FUNCTION {
-		return vm.call(callee.Obj.(*value.ObjClosure), argCount, c, ip)
-	}
-	if callee.Type == value.VAL_NATIVE {
-		native := callee.Obj.(*value.ObjNative)
-		args := vm.stack[vm.stackTop-argCount : vm.stackTop]
-		if native.Signature != nil {
-			sig := native.Signature
-			if !sig.Variadic && argCount != sig.Arity {
-				return false, vm.runtimeError(c, ip, "native '%s' expects %d arguments, got %d", native.Name, sig.Arity, argCount)
-			}
-			if sig.Variadic && argCount < sig.Arity {
-				return false, vm.runtimeError(c, ip, "native '%s' expects at least %d arguments, got %d", native.Name, sig.Arity, argCount)
-			}
-
-			params := sig.Params
-			if sig.Variadic && len(params) > 0 && argCount > len(params) {
-				expanded := make([]value.ParamInfo, argCount)
-				copy(expanded, params)
-				for i := len(params); i < argCount; i++ {
-					expanded[i] = params[len(params)-1]
-				}
-				params = expanded
-			}
-			if err := validateParameterModes(native.Name, params, args); err != nil {
-				return false, vm.runtimeError(c, ip, "%s", err)
-			}
-
-			callArgs := make([]value.Value, len(args))
-			copy(callArgs, args)
-			for i, param := range params {
-				if i >= len(callArgs) {
-					break
-				}
-				if !param.IsRef {
-					callArgs[i] = vm.copyValue(callArgs[i])
-				}
-			}
-			args = callArgs
-		}
-		// fmt.Printf("Calling native %s with args: %v\n", native.Name, args)
-		result := native.Fn(args)
-		vm.stackTop -= argCount + 1 // args + function
-		vm.push(result)
-		return true, nil
-	}
-	return false, vm.runtimeError(c, ip, "can only call functions and classes")
-}
-
-func (vm *VM) call(closure *value.ObjClosure, argCount int, c *chunk.Chunk, ip int) (bool, error) {
-	// fmt.Printf("Calling function %s, code len: %d\n", fn.Name, len(chunk.Code))
-
-	fn := closure.Function
-
-	if argCount != fn.Arity {
-		return false, vm.runtimeError(c, ip, "expected %d arguments but got %d", fn.Arity, argCount)
-	}
-
-	if vm.frameCount == FramesMax {
-		return false, vm.runtimeError(c, ip, "stack overflow")
-	}
-
-	// Handle Pass-by-Value (Copy) for non-ref parameters
-	// Args are at vm.stackTop - argCount
-	baseArgs := vm.stackTop - argCount
-	args := vm.stack[baseArgs:vm.stackTop]
-	if err := validateParameterModes(fn.Name, fn.Params, args); err != nil {
-		return false, vm.runtimeError(c, ip, "%s", err)
-	}
-	for i := 0; i < argCount; i++ {
-		if i < len(fn.Params) {
-			param := fn.Params[i]
-			if !param.IsRef {
-				// Pass by Value: Copy if mutable object
-				val := vm.stack[baseArgs+i]
-				vm.stack[baseArgs+i] = vm.copyValue(val)
-			}
-		}
-	}
-
-	frame := &CallFrame{
-		Closure: closure,
-		IP:      0,
-		Slots:   vm.stackTop - argCount - 1, // Start of locals window (fn + args)
-		Globals: closure.Globals,
-	}
-	// Push new frame
-	vm.frames[vm.frameCount] = frame
-	vm.frameCount++
-	vm.currentFrame = frame
-	return true, nil
-}
-
-func (vm *VM) copyValue(v value.Value) value.Value {
-	if v.Type != value.VAL_OBJ {
-		return v
-	}
-	switch obj := v.Obj.(type) {
-	case *value.ObjArray:
-		newElems := make([]value.Value, len(obj.Elements))
-		copy(newElems, obj.Elements)
-		copied := value.NewArray(newElems)
-		copied.Obj.(*value.ObjArray).RuntimeType.Store(obj.RuntimeType.Load())
-		return copied
-	case *value.ObjMap:
-		newData := make(map[interface{}]value.Value)
-		for k, val := range obj.Data {
-			newData[k] = val
-		}
-		copied := value.NewMap()
-		copiedMap := copied.Obj.(*value.ObjMap)
-		copiedMap.Data = newData
-		copiedMap.RuntimeType.Store(obj.RuntimeType.Load())
-		return copied
-	case *value.ObjInstance:
-		newFields := make(map[string]value.Value)
-		for k, val := range obj.Fields {
-			newFields[k] = val
-		}
-		return value.Value{Type: value.VAL_OBJ, Obj: &value.ObjInstance{Struct: obj.Struct, Fields: newFields}}
-	default:
-		return v
-	}
-}
-
-func (vm *VM) readShort() uint16 {
-	vm.ip += 2
-	return uint16(vm.chunk.Code[vm.ip-2])<<8 | uint16(vm.chunk.Code[vm.ip-1])
-}
-
-// isFalsey returns true if the value is false or null
-func isFalsey(v value.Value) bool {
-	return v.Type == value.VAL_NULL || (v.Type == value.VAL_BOOL && !v.AsBool)
-}
-
-func valuesEqual(a, b value.Value) bool {
-	if a.Type == b.Type {
-		switch a.Type {
-		case value.VAL_BOOL:
-			return a.AsBool == b.AsBool
-		case value.VAL_NULL:
-			return true
-		case value.VAL_INT:
-			return a.AsInt == b.AsInt
-		case value.VAL_FLOAT:
-			return a.AsFloat == b.AsFloat
-		case value.VAL_OBJ:
-			return a.Obj == b.Obj // Simple pointer/string comparison
-		case value.VAL_BYTES:
-			return a.Obj.(string) == b.Obj.(string)
-		default:
-			return false
-		}
-	}
-
-	// Mixed types
-	if a.Type == value.VAL_INT && b.Type == value.VAL_FLOAT {
-		return float64(a.AsInt) == b.AsFloat
-	}
-	if a.Type == value.VAL_FLOAT && b.Type == value.VAL_INT {
-		return a.AsFloat == float64(b.AsInt)
-	}
-
-	return false
-}
-
-func (vm *VM) readConstant() value.Value {
-	// Assumes 1 byte operand for constant index
-	index := vm.chunk.Code[vm.ip]
-	vm.ip++
-	return vm.chunk.Constants[index]
-}
-
-func (vm *VM) push(v value.Value) {
-	if vm.stackTop >= StackMax {
-		panic("Stack overflow")
-	}
-	vm.stack[vm.stackTop] = v
-	vm.stackTop++
-}
-
-func (vm *VM) pop() value.Value {
-	vm.stackTop--
-	val := vm.stack[vm.stackTop]
-	vm.stack[vm.stackTop] = value.Value{} // Clear reference to help GC
-	return val
-}
-
 func (vm *VM) loadModule(name string) (value.Value, error) {
 	// Convert dot notation to path path separator
 	pathName := strings.ReplaceAll(name, ".", string(filepath.Separator))
@@ -1864,53 +1647,4 @@ FileImport:
 
 	// Return the Module Map
 	return value.NewMapWithData(moduleGlobals), nil
-}
-
-func (vm *VM) peek(distance int) value.Value {
-	return vm.stack[vm.stackTop-1-distance]
-}
-
-// captureUpvalue finds or creates an open upvalue for the given stack slot.
-func (vm *VM) captureUpvalue(local *value.Value) *value.ObjUpvalue {
-	// var prevUpvalue *value.ObjUpvalue // Unused for now
-	upvalue := vm.openUpvalues
-
-	// Walk list
-	for upvalue != nil && upvalue.Location != local {
-		// prevUpvalue = upvalue
-		upvalue = upvalue.Next
-	}
-
-	if upvalue != nil && upvalue.Location == local {
-		return upvalue
-	}
-
-	createdUpvalue := &value.ObjUpvalue{
-		Location: local,
-		Next:     vm.openUpvalues,
-	}
-	vm.openUpvalues = createdUpvalue
-
-	return createdUpvalue
-}
-
-func (vm *VM) closeUpvalue(slot *value.Value) {
-	var prev *value.ObjUpvalue
-	curr := vm.openUpvalues
-
-	for curr != nil {
-		if curr.Location == slot {
-			curr.Closed = *slot          // Copy value to heap
-			curr.Location = &curr.Closed // Point to heap
-
-			if prev == nil {
-				vm.openUpvalues = curr.Next
-			} else {
-				prev.Next = curr.Next
-			}
-			return
-		}
-		prev = curr
-		curr = curr.Next
-	}
 }
